review: int64 variance accumulator, int32 silent_ms, starved-calibration timeout
- compute_amp_variance's per-subcarrier sum-of-squares goes int64: the
  normalized amplitude can reach 64*N (~7.3k at HT40's 114 subcarriers)
  when one subcarrier dominates, and 40 frames of a^2 brushes INT32_MAX
  (UB on overflow). All three copies, sync guards clean.
- supply.silent_ms serialized via (int32_t) cast + %d — the UINT32_MAX
  sentinel prints -1 portably.
- calibration timeout accounting split out of calibration_observe and
  ticked before the honesty gate, so a frame-starved install reports
  calibrate status timed_out instead of running forever (Codex P2).

+++ firmware/canary/lib/securacv_csi/src/securacv_csi.cpp
@@ -322,11 +322,11 @@ namespace features {
     const int32_t F = (int32_t)s_frame_count;
     for (size_t k = 0; k < s_sc_count; k++) {
       int32_t sum = 0;
-      int32_t sq  = 0;
+      int64_t sq  = 0;   /* a ≤ 64·N ⇒ 40·a² brushes INT32_MAX at HT40 */
       for (size_t f = 0; f < s_frame_count; f++) {
         const int32_t a = s_amp_hist[f][k];
         sum += a;
-        sq  += a * a;
+        sq  += (int64_t)a * a;
       }
       /* Single-step variance — no truncated-mean bias. */
       int32_t var = (int32_t)((sq - ((int64_t)sum * sum) / F) / F);
